requirements.txt
backcall==0.1.0
bleach==3.1.0
chess-py==3.3.1
colorama==0.4.1
cycler==0.10.0
decorator==4.3.0
defusedxml==0.5.0
entrypoints==0.3
ipykernel==5.1.0
ipython==7.2.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
jedi==0.13.2
Jinja2==2.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
kiwisolver==1.0.1
MarkupSafe==1.1.0
matplotlib==3.0.2
mistune==0.8.4
nbconvert==5.4.0
nbformat==4.4.0
notebook==5.7.4
numpy==1.15.4
pandas==0.23.4
pandoc==1.0.2
pandocfilters==1.4.2
parso==0.3.1
pickleshare==0.7.5
ply==3.11
prometheus-client==0.5.0
prompt-toolkit==2.0.7
Pygments==2.3.1
pyparsing==2.3.1
python-chess==0.24.1
python-dateutil==2.7.5
pytz==2018.9
pywinpty==0.5.5
pyzmq==17.1.2
qtconsole==4.4.3
Send2Trash==1.5.0
simplegeneric==0.8.1
six==1.12.0
terminado==0.8.1
testpath==0.4.2
tornado==5.1.1
traitlets==4.3.2
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.4.2
